+++ src/Router.jsx
@@ -3,10 +3,9 @@ import Home from './pages/Home';
 import Maps from './pages/Maps';
 import SignUp from './pages/SignUp';
 import Login from './pages/Login';
-import Layout from './components/Layout';
+import Layout from './components/layout/Layout';
 import Bookmark from './pages/Bookmark';
-import ProtectedRoute from './components/ProtectedRoute';
-import SharePreview from './pages/SharePreview'; // 공용컴포넌트 미리보는 페이지
+import ProtectedRoute from './components/middleware/ProtectedRoute';
 
 const Router = () => {
   return (
@@ -14,7 +13,6 @@ const Router = () => {
       <Routes>
         <Route path="/" element={<Layout />}>
           <Route index element={<Home />} />
-          <Route path="/share" element={<SharePreview />} />
           <Route path="/login" element={<Login />} />
           <Route path="/signup" element={<SignUp />} />
           <Route path="/map" element={<Maps />} />
