+++ src/components/bookmark/ProfileContainer.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import Button from '../Button';
+import Button from '../common/Button';
 import { useUserStore } from '../../zustand/userStore';
 
 const ProfileContainer = ({ openModal }) => {
