+++ src/components/common/DetailContent.jsx
